uploaded images are displayed as they are uploaded
