+++ frontend/src/Components/Profile/ProfileDetailsCard.js
@@ -40,42 +40,14 @@ const useStyles = makeStyles((theme) => ({
   },
 }));
 
-const ProfileDetailsCard = ({
-  editMode,
-  editImage,
-  formFields,
-  profileDetails,
-  selectedImage,
-  setSelectedImage,
-  uploadedImage,
-  setUploadedImage,
-  handleRemove,
-  submit,
-  avatarSrc,
-  toggleEditImage,
-  handleInput,
-  handleSubmit,
-  handleToggle,
-}) => {
+const ProfileDetailsCard = ({ editMode, formFields, handleInput, handleSubmit, handleToggle, profileDetails }) => {
   const classes = useStyles();
 
   return (
     <Card className={classes.root}>
       <CardContent>
         <Box className={classes.rowContainer}>
-          <UserProfile
-            editImage={editImage}
-            formFields={formFields}
-            avatarSrc={avatarSrc}
-            selectedImage={selectedImage}
-            setSelectedImage={setSelectedImage}
-            uploadedImage={uploadedImage}
-            setUploadedImage={setUploadedImage}
-            handleRemove={handleRemove}
-            submit={submit}
-            profileDetails={profileDetails}
-            toggleEditImage={toggleEditImage}
-          />
+          <UserProfile formFields={formFields} profileDetails={profileDetails} />
           <Box className={classes.emptyGap}></Box>
           <Box>
             <IconButton disabled={!editMode} onClick={handleSubmit} className={classes.buttonSecondary}>
